version 3.30.15.10 1. Implemented selectivity curve 43 - scaled linear segments, curve 44 -    two sex random walk, and curve 45 - two sex logist values. 2. Changed reading function to allow breaking out of reading bad files. This    leaves SSI in an unstable state and must be followed by reading a good    model. 3. Fixed curve 16 - Coleraine - Gaussian: limit for selex = 1.

+++ src/population_widget.cpp
@@ -902,7 +902,7 @@ void population_widget::changeGrowthBinMethod(int value)
 
     }
 }
-void population_widget::changeGrowthBinMin(float min)
+void population_widget::changeGrowthBinMin(double min)
 {
     pop->Grow()->setGrowthBinMin(min);
     if (pop->Grow()->getGrowthBinMethod() == 2) {
@@ -911,7 +911,7 @@ void population_widget::changeGrowthBinMin(float min)
         growthBinsView->resizeColumnsToContents();
     }
 }
-void population_widget::changeGrowthBinMax(float max)
+void population_widget::changeGrowthBinMax(double max)
 {
     pop->Grow()->setGrowthBinMax(max);
     if (pop->Grow()->getGrowthBinMethod() == 2) {
@@ -920,7 +920,7 @@ void population_widget::changeGrowthBinMax(float max)
         growthBinsView->resizeColumnsToContents();
     }
 }
-void population_widget::changeGrowthBinWidth(float wid)
+void population_widget::changeGrowthBinWidth(double wid)
 {
     pop->Grow()->setGrowthBinStep(wid);
     if (pop->Grow()->getGrowthBinMethod() == 2) {
